ParticipantsTable unit tests
- Moved tests to separate file
- Added test for closed request state
